requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3; python_version >= "3.5"
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
django==3.1.3
django-cockroachdb==3.1.2
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.11.2
drf-yasg==1.20.0
factory-boy==2.12.0
faker==4.17.1
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0; python_version >= "3.5"
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.13.0
-e git+https://github.com/FNSdev/graphene-django-optimizer.git@e1a3d0412d06f6b95bb9505d5fa6f161c7013446#egg=graphene-django-optimizer
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
idna==2.10
inflection==0.5.1
itypes==1.2.0
jinja2==2.11.2
markupsafe==1.1.1
packaging==20.4
promise==2.3
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20; python_version >= "3.5"
pyjwt==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.4
requests==2.25.0
rsa==4.6; python_version >= "3.5"
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2; platform_python_implementation == "CPython" and python_version < "3.9"
rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.11.8
uvloop==0.14.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
websockets==8.1
